Split git and path utilities into packages
