+++ change/workspace-tools-63ff6dc4-8f44-4e2d-be6a-6d10c7cf09cb.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Move git and path utilities into separate packages (re-exported by this package)",
+  "packageName": "workspace-tools",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
