+++ change/workspace-tools-git-7d7fbf7e-ec0f-411d-99cc-0d240742785a.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Create package",
+  "packageName": "workspace-tools-git",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
