+++ change/workspace-tools-paths-a95113b6-5cb6-4d05-ac86-f6195863decf.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Create package",
+  "packageName": "workspace-tools-paths",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
