+++ packages/workspace-tools-git/README.md
@@ -1,0 +1,9 @@
+# workspace-tools-git
+
+Git-related helper methods.
+
+## Environment Variables
+
+### GIT_MAX_BUFFER
+
+Override the `maxBuffer` value for git processes, for example if the repo is very large. `workspace-tools` uses 500MB by default.
