+++ packages/workspace-tools-git/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("workspace-tools-scripts/jest/jest.config");
