+++ packages/workspace-tools-git/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "workspace-tools",
-  "version": "0.27.0",
+  "name": "workspace-tools-git",
+  "version": "0.1.0",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -17,12 +17,7 @@
     "test": "jest"
   },
   "dependencies": {
-    "@yarnpkg/lockfile": "^1.1.0",
-    "globby": "^11.0.0",
-    "jju": "^1.4.0",
-    "js-yaml": "^4.1.0",
-    "micromatch": "^4.0.0",
-    "workspace-tools-git": "^0.1.0",
+    "git-url-parse": "^13.0.0",
     "workspace-tools-paths": "^0.1.0"
   },
   "devDependencies": {
