+++ packages/workspace-tools-paths/README.md
@@ -1,0 +1,3 @@
+# workspace-tools-paths
+
+Helper methods related to paths in a repo or monorepo.
