+++ packages/workspace-tools-paths/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "workspace-tools",
-  "version": "0.27.0",
+  "name": "workspace-tools-paths",
+  "version": "0.1.0",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -15,15 +15,6 @@
     "build": "tsc",
     "start": "tsc -w --preserveWatchOutput",
     "test": "jest"
-  },
-  "dependencies": {
-    "@yarnpkg/lockfile": "^1.1.0",
-    "globby": "^11.0.0",
-    "jju": "^1.4.0",
-    "js-yaml": "^4.1.0",
-    "micromatch": "^4.0.0",
-    "workspace-tools-git": "^0.1.0",
-    "workspace-tools-paths": "^0.1.0"
   },
   "devDependencies": {
     "workspace-tools-scripts": "*"
