+++ packages/workspace-tools-paths/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "workspace-tools-scripts/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "lib"
+  },
+  "include": ["src"]
+}
